+++ flutter-hms-location/CHANGELOG.md
@@ -1,6 +1,7 @@
-## 5.0.0+301
+## 5.0.0+301 
 
-* Upgrade HMS Location Android SDK to 5.0.0.301.
+* Added the getNavigationContextState function.
+* Added the enableLogger and disableLogger methods.
 
 ## 4.0.4+1/4.0.4+300