+++ flutter-hms-site/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 5.0.1+300
+
+* Added the **siteSearchActivity** method.
+* Added the **SearchIntent** class to start site search activity.
+* Added the **SearchFilter** class, which is used in a SearchIntent object as **filter**.
+* Added the **streetNumber** attribute in the **AddressDetail** class, which is used for setting or obtaining the street number of a place.
+* Added the **HwLocationType** class, which contains constants of Huawei POI types.
+* Added the **enableLogger** and **disableLogger** methods.
+
 ## 4.0.4
 
 * Initial release.